Convertendo algumas operações para JPA e corrigindo um problema decorrente do banco de homologação ser case-sensitive

+++ src/main/java/br/jus/trf2/balcaovirtual/model/CnjClasse.java
@@ -4,37 +4,36 @@ import java.io.Serializable;
 import javax.persistence.*;
 import java.util.List;
 
-
 /**
  * The persistent class for the cnj_classe database table.
  * 
  */
 @Entity
-@Table(name="cnj_classe")
-@NamedQuery(name="CnjClasse.findAll", query="SELECT c FROM CnjClasse c")
+@Table(name = "cnj_classe")
+@NamedQuery(name = "CnjClasse.findAll", query = "SELECT c FROM CnjClasse c")
 public class CnjClasse implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	@Column(name="CNCL_ID")
-	private int cnclId;
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "CNCL_ID")
+	private Long cnclId;
 
-	@Column(name="CNCL_NM")
+	@Column(name = "CNCL_NM")
 	private String cnclNm;
 
-	//bi-directional many-to-many association to TipoMarca
-	@ManyToMany(mappedBy="cnjClasses")
+	// bi-directional many-to-many association to TipoMarca
+	@ManyToMany(mappedBy = "cnjClasses")
 	private List<TipoMarca> tipoMarcas;
 
 	public CnjClasse() {
 	}
 
-	public int getCnclId() {
+	public Long getCnclId() {
 		return this.cnclId;
 	}
 
-	public void setCnclId(int cnclId) {
+	public void setCnclId(Long cnclId) {
 		this.cnclId = cnclId;
 	}
